Working state (again).
Cleared most warnings by changing a couple of access modifiers.

Previously I only retrieved a thumbnail, now I only retrieve the full image.
It required some time because the authorization must end with .fileprovider for this to work and it is not possible to broadcast it to the Gallery since the file was created with GetExternalFilesDir().

I should retrieve the image through the Room-stored imageUri to show it in the RecycleView.
For this to work, the RecycleView should not use simple_list_item1 any longer since that one only contains a TextView. Instead, I should create a new custom view to store the image and/ or text.